feat: refactor authentication and user validation in employee dashboard and holidays endpoints to use Users model, improve error handling, and ensure consistent payroll data retrieval and refactor attendance management to utilize Users model, enhance error handling, and streamline check-in/check-out processes
